refactor: Melhorando os DTOs, Entidades, Controllers e Services, deixando mais legíveis e melhorando os métodos.

+++ src/main/java/fabriciocarvalho348/com/github/vendasapi/dto/ItemPedidoDto.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ItemPedidoDto {
 
-    private Integer produto;
+    private Long produto;
     private Integer quantidade;
 }
